Bug fix for PageNavigation with template and absolute uri
This is a bug fix for when the PageNavigation attribute is being used
with a template and absolute uri is set to true. The navigation should
go to your specified url, but prior to this fix, would still preappend
the base root onto the absolute uri because the absolute uri property
was getting ignored.